backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg==0.29.0
pydantic>=2.5.0
anyio>=4.5.0

# ML (gleiche Versionen wie Training Service!)
scikit-learn==1.3.2
xgboost==2.0.2
  # Für SMOTE (Imbalanced Data Handling)
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Utilities
prometheus-client==0.19.0
python-dateutil==2.8.2
aiohttp>=3.9.1  # Für n8n Webhooks
httpx>=0.25.2  # Für HTTP-Requests in Streamlit
requests==2.31.0  # Für synchrone HTTP-Requests
streamlit==1.28.0  # Für Web UI
plotly==5.18.0  # Für Charts in Streamlit

# Logging (optional, für strukturiertes Logging)
python-json-logger==2.0.7  # Optional, wenn JSON-Logging gewünscht

# MCP Server (Model Context Protocol)
mcp>=1.0.0
sse-starlette>=1.6.5

